sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hatidata-agent"
version = "0.1.0"
description = "HatiData Agent SDK — RAM for Agents"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "Marviy Pte Ltd", email = "eng@hatidata.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
async = ["asyncpg>=0.29"]
langchain = ["langchain-community>=0.2"]
mcp = ["mcp>=1.0"]
all = [
    "asyncpg>=0.29",
    "langchain-community>=0.2",
    "mcp>=1.0",
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]

[project.urls]
Homepage = "https://hatidata.com"
Documentation = "https://docs.hatiosai.com/hatidata"
Repository = "https://github.com/HatiOS-AI/hatidata-core"
Changelog = "https://github.com/HatiOS-AI/hatidata-core/releases"

[project.scripts]
hatidata-mcp-server = "hatidata_agent.mcp_server:main"

[tool.setuptools.packages.find]
where = ["."]
